// typings.d.ts
type Base = {
    _createdAt: string;
    _id: string;
    _rev: string;
    _type: string;
    _updatedAt: string;
    _publishedAt: string;
};
// ...
interface TeamMember extends Base{
    name: string;
    position: string;
    image: string;
    slug: slug,
    socials: Socials[];
    teamDetails: TeamDetails[]
}
interface SkillTitle {
    title?: string;
    percentage?: string;
}
interface TeamDetails  {
    image:  string,
    phone: string,
    email: string;
    location: string;
    specialization: string;
    name: string;
    bio: string;
    socials: Socials[],
    skills: SkillTitle[],
    skillSetInfo: string;
}
// ...
interface Socials {
   
    // add more social media platforms as needed
    platform: 'twitter' | 'instagram' | 'facebook' | 'linkedin' | 'youtube' | 'twitch';
    url: string;
}

// TEAM 
interface TeamProps{
    name: string;
    position: string;
    image: string;
    slug: slug,
    socials: Socials[];
    // teamDetails: TeamDetails[]
}